+++ datalake/.minio.sys/format.json
@@ -1,0 +1,1 @@
+{"version":"1","format":"xl-single","id":"dbd9b957-3d5d-4dd7-9e27-7b5a038f56c6","xl":{"version":"3","this":"5abaab11-6959-4581-bd49-aa0c8d60590c","sets":[["5abaab11-6959-4581-bd49-aa0c8d60590c"]],"distributionAlgo":"SIPMOD+PARITY"}}
