+++ pyproject.toml
@@ -1,0 +1,17 @@
+[project]
+name = "paris-events-analyzer"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "duckdb>=1.3.1",
+    "minio>=7.2.15",
+    "requests-cache>=1.2.1",
+    "rich>=14.0.0",
+]
+
+[dependency-groups]
+dev = [
+    "rust-just>=1.40.0",
+]
