--- requirements.in
@@ -1,4 +1,0 @@
-requests-cache
-duckdb
-minio
-rich
