+++ src/ingestion/endpoints.json
@@ -1,0 +1,5 @@
+{
+  "csv": "https://opendata.paris.fr/api/explore/v2.1/catalog/datasets/que-faire-a-paris-/exports/csv",
+  "json": "https://opendata.paris.fr/api/explore/v2.1/catalog/datasets/que-faire-a-paris-/exports/json",
+  "parquet": "https://opendata.paris.fr/api/explore/v2.1/catalog/datasets/que-faire-a-paris-/exports/parquet"
+}
